Handle opening of diagrams in main window
